Add modified SHMI  weights
SHMI weighting for diabetes with complications is -1. However, when using SHMI Charlson  score in isolation as a measure of comorbidity, it does not make sense for diabetes with complications to carry a smaller weight than diabetes let alone a negative weight.

Hence, a pragmatic weight of 4 is assigned. Weight for diabetes is 3.